add adjustable settings file for SQL connection and window size/position

+++ src/ShoppingListApp.java
@@ -1,6 +1,7 @@
 import fi.tamk.tiko.MyListPackage.MyLinkedList;
 
 import java.io.*;
+import java.util.Properties;
 
 /**
  * ShoppingListApp
@@ -10,22 +11,107 @@ import java.io.*;
  * @since       1.8
  */
 public class ShoppingListApp {
+    private Properties settings;
     private MyLinkedList<ShoppingItem> shoppingList;
     private SQLmanager sqlManager;
+    private File myFolder;
+    private File mySettingsFile;
+    private boolean online;
+    private final String HOST_TAG = "host";
+    private final String DATABASE_TAG = "database";
+    private final String USERNAME_TAG = "username";
+    private final String PASSWORD_TAG = "password";
 
     /**
      * Instantiates a new ShoppingListApp.
      */
     public ShoppingListApp() {
+        setOnline(true);
+        localServices();
+        runWebServices();
+    }
+
+    public ShoppingListApp(boolean online) {
+        setOnline(online);
+        localServices();
+
+        if (isOnline()) {
+            runWebServices();
+        }
+    }
+
+    private void localServices() {
+        String myFolderPath = System.getProperty("user.home")
+                                + File.separator
+                                + "MasterShopper9000";
+        String mySettingsFilePath = myFolderPath
+                                    + File.separator
+                                    + "settings.cfg";
+        myFolder = new File(myFolderPath);
+        mySettingsFile = new File(mySettingsFilePath);
+        settings = new Properties();
         shoppingList = new MyLinkedList<>();
-        sqlManager = new SQLmanager();
+
+        if (!myFolder.exists()) {
+            myFolder.mkdirs();
+        }
+
+        if (!mySettingsFile.exists()) {
+            setDefaultSQLSettings();
+        }
+    }
+
+    public void runWebServices() {
+        if (sqlManager != null) {
+            sqlManager.close();
+            sqlManager = null;
+        }
+
+        try (InputStream input = new FileInputStream(mySettingsFile)) {
+            settings.load(input);
+            sqlManager = new SQLmanager(settings.getProperty(HOST_TAG),
+                                        settings.getProperty(DATABASE_TAG),
+                                        settings.getProperty(USERNAME_TAG),
+                                        settings.getProperty(PASSWORD_TAG));
+        } catch (IOException e) {
+            e.printStackTrace();
+            setOnline(false);
+        }
+    }
+
+    public void setDefaultSQLSettings() {
+        try (OutputStream output = new FileOutputStream(mySettingsFile)) {
+            settings.setProperty(HOST_TAG, "mydb.tamk.fi");
+            settings.setProperty(DATABASE_TAG, "dbc6jyrvir1");
+            settings.setProperty(USERNAME_TAG, "c6jyrvir");
+            settings.setProperty(PASSWORD_TAG, "1Salasana");
+            settings.store(output, "Settings file for ShoppingListApp");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void setCustomSQLSettings(String host, String database,
+                                  String userName, String password) {
+
+        try (OutputStream output = new FileOutputStream(mySettingsFile)) {
+            settings.setProperty(HOST_TAG, host);
+            settings.setProperty(DATABASE_TAG, database);
+            settings.setProperty(USERNAME_TAG, userName);
+            settings.setProperty(PASSWORD_TAG, password);
+            settings.store(output, "Settings file for ShoppingListApp");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
     /**
      * Shutdown procedures.
      */
     public void close() {
-        sqlManager.close();
+        if (isOnline()) {
+            sqlManager.close();
+        }
     }
 
     /**
@@ -212,14 +298,56 @@ public class ShoppingListApp {
      * Saves shoppingList contents to SQL database.
      */
     public void saveToSQL() {
-        sqlManager.deleteAll();
-        sqlManager.save(getShoppingList());
+        if (isOnline()) {
+            sqlManager.deleteAll();
+            sqlManager.save(getShoppingList());
+        }
+    }
+
+    public void addNewSetting(String tag, String value) {
+        try (OutputStream out = new FileOutputStream(mySettingsFile)) {
+            settings.setProperty(tag, value);
+            settings.store(out, "Settings file for ShoppingListApp");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public String readSetting(String tag) {
+        String result = "";
+
+        try (InputStream input = new FileInputStream(mySettingsFile)) {
+            settings.load(input);
+            result = settings.getProperty(tag);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        return result;
     }
 
     /**
      * Loads shoppingList contents from SQL database.
      */
     public void loadFromSQL() {
-        setShoppingList(sqlManager.load());
+        if (isOnline()) {
+            setShoppingList(sqlManager.load());
+        }
+    }
+
+    public File getMyFolder() {
+        return myFolder;
+    }
+
+    public File getMySettingsFile() {
+        return mySettingsFile;
+    }
+
+    public boolean isOnline() {
+        return online;
+    }
+
+    public void setOnline(boolean online) {
+        this.online = online;
     }
 }
